requirements.txt
Flask
Flask-RESTful
uwsgi
psycopg2
pillow
